# openapi.yaml
openapi: 3.0.1
info:
  title: Mini Search Engine API
  version: 0.1.0
paths:
...
  /api/image/query:
    get:
      summary: get list of urls to images that are associated with given search terms
      operationId: imageQuery
      parameters:
        - name: q
          in: query
          description: string to search for
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: page of results to fetch, default 1
          schema:
            type: integer
        - name: country 
          in: query
          description: country alpha-3 code (ISO 3166) all capital
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation, result mustn't be empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultPage"
        "404":
          description: no results
...
components:
  schemas:
    ResultPage:
      type: object
      title: ResultPage
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/QueryResult"
    QueryResult:
      type: object
      title: QueryResult
      properties:
        title:
          type: string
        link:
          type: string
        snippet:
          type: string
...